forms-flow-data-analysis-api/requirements/prod.txt
gunicorn
Flask
Flask-Caching
Flask-SQLAlchemy
flask-restx
flask-marshmallow==0.11.0
flask-jwt-oidc
python-dotenv
psycopg2-binary
marshmallow-sqlalchemy
requests
attrs
Werkzeug
sqlalchemy_utils
sqlalchemy
spacy
flask-cors
Flask-PyMongo
nltk
Flask-Migrate
Flask-Script
